Adding (hopefully) all Dwarven Mines commissions

+++ src/main/java/io/github/moulberry/notenoughupdates/overlays/MiningOverlay.java
@@ -491,38 +491,59 @@ public class MiningOverlay extends TextTabOverlay {
 	}
 
 	private String getTip(String name) {
-		if (name.equals("Chest Looter")) return "Open 3 chests";
-		if (name.equals("Hard Stone Miner")) return "Break 1,000 Hard Stone";
+		if (SBInfo.getInstance().getLocation().equals("mining_3")) { // Dwarven Mines
+			if (name.equals("Lucky Raffle")) return "Collect 20 Raffle Tickets during §6Raffle event";
+			if (name.equals("Goblin Raid Slayer")) return "Kill 20 Goblins during §6Goblin Raid event";
 
-		String jungle = " §a(Jungle)";
-		String goblin = " §6(Golbin Holdout)";
-		String mithril = " §b(Mithril Deposits)";
-		String precursor = " §7(Precursor Remenants)";
-		String magma = " §c(Magma Fields)";
+			if (name.equals("Mithril Miner")) return "Break 500 Mithril (everywhere)";
+			if (name.equals("Titanium Miner")) return "Break 15 Titanium (everywhere)";
+			if (name.equals("Ice Walker Slayer")) return "Kill 50 Ice Walkers §b(Great Ice Wall)";
+			if (name.equals("Goblin Slayer")) return "Kill 100 Goblins §b(Goblin Borrows)";
 
-		if (name.equals("Goblin Slayer")) return "Kill 13 Goblins" + goblin;
-		if (name.equals("Sludge Slayer")) return "Kill 25 Sludges" + jungle;
-		if (name.equals("Thyst Slayer")) return "Kill 5 Thysts, when breaking Amethysts" + jungle;
-		if (name.equals("Boss Corleone Slayer")) return "Find and kill Corleone" + mithril;
-		if (name.equals("Yog Slayer")) return "Kill 13 Yogs" + magma;
-		if (name.equals("Automaton Slayer")) return "Kill 13 Automatons" + precursor;
-		if (name.equals("Team Treasurite Member Slayer")) return "Kill 13 Team Treasurite Members" + mithril;
+			if (name.equals("Cliffside Veins Mithril")) return "Break 350 Mithril §b(Cliffside Veins)";
+			if (name.equals("Royal Mines Mithril")) return "Break 350 Mithril §b(Royal Mines)";
+			if (name.equals("Lava Springs Mithril")) return "Break 350 Mithril §b(Lava Springs)";
+			if (name.equals("Rampart's Quarry Mithril")) return "Break 350 Mithril §b(Rampart's Quarry)";
 
-		if (name.endsWith("Crystal Hunter")) {
-			if (name.startsWith("Amethyst")) return "Temple Jump & Run" + jungle;
-			if (name.startsWith("Jade")) return "4 weapons from Mines of Divan" + mithril;
-			if (name.startsWith("Amber")) return "King and Queen" + goblin;
-			if (name.startsWith("Sapphire")) return "6 Robot Parts in Precursor City" + precursor;
-			if (name.startsWith("Topaz")) return "Kill Bal" + magma;
-		}
+			if (name.equals("Cliffside Veins Titanium")) return "Break 10 Titanium §b(Cliffside Veins)";
+			if (name.equals("Lava Springs Titanium")) return "Break 10 Titanium §b(Lava Springs)";
+			if (name.equals("Royal Mines Titanium")) return "Break 10 Titanium §b(Royal Mines)";
+			if (name.equals("Rampart's Quarry Titanium")) return "Break 10 Titanium §b(Rampart's Quarry)";
 
-		if (name.endsWith("Gemstone Collector")) {
-			if (name.startsWith("Amber")) return "Break orange" + goblin;
-			if (name.startsWith("Sapphire")) return "Break blue" + precursor;
-			if (name.startsWith("Jade")) return "Break green" + mithril;
-			if (name.startsWith("Amethyst")) return "Break purple" + jungle;
-			if (name.startsWith("Ruby")) return "Break red (everywhere)";
-			if (name.startsWith("Topaz")) return "Break yellow" + magma;
+		} else if (SBInfo.getInstance().getLocation().equals("crystal_hollows")) { // Crystal Hollows
+			if (name.equals("Chest Looter")) return "Open 3 chests";
+			if (name.equals("Hard Stone Miner")) return "Break 1,000 Hard Stone";
+
+			String jungle = " §a(Jungle)";
+			String goblin = " §6(Golbin Holdout)";
+			String mithril = " §b(Mithril Deposits)";
+			String precursor = " §7(Precursor Remenants)";
+			String magma = " §c(Magma Fields)";
+
+			if (name.equals("Goblin Slayer")) return "Kill 13 Goblins" + goblin;
+			if (name.equals("Sludge Slayer")) return "Kill 25 Sludges" + jungle;
+			if (name.equals("Thyst Slayer")) return "Kill 5 Thysts, when breaking Amethysts" + jungle;
+			if (name.equals("Boss Corleone Slayer")) return "Find and kill Corleone" + mithril;
+			if (name.equals("Yog Slayer")) return "Kill 13 Yogs" + magma;
+			if (name.equals("Automaton Slayer")) return "Kill 13 Automatons" + precursor;
+			if (name.equals("Team Treasurite Member Slayer")) return "Kill 13 Team Treasurite Members" + mithril;
+
+			if (name.endsWith("Crystal Hunter")) {
+				if (name.startsWith("Amethyst")) return "Temple Jump & Run" + jungle;
+				if (name.startsWith("Jade")) return "4 weapons from Mines of Divan" + mithril;
+				if (name.startsWith("Amber")) return "King and Queen" + goblin;
+				if (name.startsWith("Sapphire")) return "6 Robot Parts in Precursor City" + precursor;
+				if (name.startsWith("Topaz")) return "Kill Bal" + magma;
+			}
+
+			if (name.endsWith("Gemstone Collector")) {
+				if (name.startsWith("Amber")) return "Break orange" + goblin;
+				if (name.startsWith("Sapphire")) return "Break blue" + precursor;
+				if (name.startsWith("Jade")) return "Break green" + mithril;
+				if (name.startsWith("Amethyst")) return "Break purple" + jungle;
+				if (name.startsWith("Ruby")) return "Break red (everywhere)";
+				if (name.startsWith("Topaz")) return "Break yellow" + magma;
+			}
 		}
 
 		return null;
